refactor: move chart helper functions to src/helpers directory
This refactoring improves the organization of the codebase by moving chart-related helper functions from the components directory to a more general helpers directory. This change enhances code reusability and maintainability by making these functions more accessible to other parts of the application. The commit also includes the addition of unit tests for the calculateCountForDataOnDates function, ensuring its reliability and correctness.
