--- packages/ui/src/components/Charts/helpers/index.ts
@@ -1,2 +1,0 @@
-export * from "./calculateCountForDataOnDates";
-export * from "./getPayloadConfigFromPayload";
