+++ packages/ui/src/helpers/getPayloadConfigFromPayload.ts
@@ -1,4 +1,4 @@
-import type { ChartConfig } from "../types";
+import type { ChartConfig } from "../components/Charts/types";
 
 function isRecord(value: unknown): value is Record<string, unknown> {
   return typeof value === "object" && value !== null && !Array.isArray(value);
